requirements.txt
fastapi
openai
pydantic
sqlalchemy
psycopg2-binary
sqlmodel
python-multipart
python-jose[cryptography]
mangum
python-dotenv
uvicorn
# Slim requirements for Vercel Serverless
